requirements.txt
requests
beautifulsoup4
lxml
shodan
dnspython
scapy
python-nmap
ipwhois
pwntools
impacket
paramiko
cryptography
pycryptodome
pexpect
netfilterqueue
yara-python
volatility3
capstone
uncompyle6
lief
r2pipe
angr
psutil
pyinotify ; platform_system == "Linux"
colorama
opencv-python
python-whois
builtwith
python-dotenv
